implement game state machine
Admin now can control the state of the game, as outlined in game_state.h

game_server does minimal work in processing commands. For a given
request the server receives, it merely sanitizes it and depending on
whether it is from the admin or a player, passes it to the game_state to
process.