GameServer uses libevent
Listening and accepting a client is handled by libevent. This means I
will now be able to easily make future events (ev_timer) for upgrading,
training, and attacking.

One concern is the order in which client connections are accepted. I
will have to think of a clever way to round-robin clients, or enforce a
rate limit on requests.

It would be good to also have the read/process/write functionality in
libevent-style (not just listening).