UPGRADE with timer events
When a city upgrades, a timer is created with the appropirate duration.
When the timer triggers its callback, the city level increases. If the
new city level is the maximum level, then another callback is triggered
on the game state, and changes the state to FINISHED.

This means ever player will be continuing play as normal, and all of a
sudden the game state will change and all requests thereafter will start
with "GAME OVER...", and indicate why the game finished (admin forced
finish, or which city one).